Implement FP exceptions in ROB
The logic here is a little confusing, because RISC-V can only raise
exceptions at 2 times (according to the ISA standard): IF & ID or
LD/ST. BOOM detects IF/ID exceptions in their respective phases and
carries that information on the uop. If an older load/store is also an
exception, some logic is needed to make sure the older
uop/instruction's exception is presented first.

We are adding another condition to this, since FP instructions can
raise exceptions during EXE phase, which in an OoO core is not too
hard to deal with. However, it is just another thing that we need to
be careful of.

fflags still accrue the same way as before, so fflags accumulating as
FP instructions are used and FP events happen remains the same.

Co-authored-by: Karl Hallsby <karl@hallsby.com>
Co-authored-by: Liam Strand <strandliam@gmail.com>