Mark FP micro-ops as unsafe instructions; do not speculate past them
There are a handful of instructions/uops that an OoO core is not
allowed to speculate past, because if a speculated instruction is
committed too early (before an earlier unsafe instruction commits) we
may alter programmer/ISA-visible state that we should not have.

FP instructions are allowed to raise exceptions now, which means that
any instruction that may depends on this FP instruction happening
could now be misspeculated and end up committing when it should not.
The easiest thing is to just mark all FP uops as unsafe for now.