Move FP exception generation to BOOM's FP execution units
This was done because the detection of an FP exception in the ROB
always ends up on the critical path in the ROB, regardless of clock
frequency of the FPGA fabric.

NOTE: The FPU logic is relatively "stateless" when it comes to FP
event flags that were raised by prior instructions/uops. This means
that each instruction run through the FPU will only raise flags for
ITSELF! We rely on the ROB to both reorder the FP instructions into the
correct order and to accrue the flags in a sticky fashion.

The BOOM's wrappers around Rocket's FP stuff and hardfloat are mostly
wires to make interacting with multiple FP execution units behave like
working with normal Scala collections. BUT, the result produced by the
FP execution units are produced ONE (1) clock cycle earlier than the
ROB sees the exception.

This is the first step in making the FP modules produce FP exceptions.
The next step is for the FP execution units to produce an FP exception
of the Exception class. Then the ROB is only responsible for
sequencing the exceptions it has coming into it, instead of building
any structures.

Lastly, below are some notes about how all of these units connect
together, which should hopefully help explain why I added things where
I did.

FPU layering (From top-down/parent-child/root-leaf):
  1. Boom Core (~core.scala~)
  2. fp_pipeline: FpPipeline (~core.scala~)
  3. FpPipeline (~fp-pipeline.scala~)
     Creates all FP modules, including the exe_units.
     exe_units: ExecutionUnits(fpu=true)
  4. exe_units: ExecutionUnits(fpu=true) (~execution-units.scala~)
     1. ExecutionUnits are a way to treat all execution units as a
        list-like structure. They allow mapping, filtering, etc. over all
        execution units based on what you want. You can then grab IOs for
        each module to do something. This means that EVERY EXE UNIT has
        the same IO interface
  5. if (!fpu) { int/mem; } else { fpu_exe_unit: FPUExeUnit } (~execution-units.scala~)
     1. Only 1 Div/sqrt FUs created, regardless of core width.
        Only 1 Fpiu FUs created, regardless of core width (a minimum
        requirement for the FRF to talk with integer).
        This is determined by for (w <- 0..fp_width) and using (w==0).
     2. You will end up with fp_width FPU units
  6. FPUExeUnit(hasFpu, hasFdiv, hasFpiu) (~execution-unit.scala~)
     When hasFpu=true, create fpu.
     fpu: FPUUnit
  7. FPUUnit (~functional-unit.scala~)
     NOTE: Bypassing unsupported!
     Extends PipelinedFunctionalUnit
     fpu: FPU
  8. FPU (~fpu.scala~)
     Wraps Boom FPU signals into Rocket's FPU.
     Rocket's FPU wraps hardfloat.
     NOTE: All FUs (fma, fpiu, fpmu) are padded to SAME latency to
     make write scheduling easier.
  9. d/sfma: FPUFMAPipe (~rocket/.../tile/FPU.scala~)
     This hooks up hardfloat's ~MulAddRecFNPipe~ (pipelined FMA with
     hardfloat's recoded format function).
  10. Hardfloat MulAddRecFNPipe